requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
python-jose==3.3.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.2
swagger-spec-validator==3.0.3
PyJWE==1.0.0
pillow==10.2.0
psycopg2-binary==2.9.9
swagger-spec-validator==3.0.3
